workers/od-training/requirements.txt
# Core ML
torch>=2.4.0
torchvision>=0.19.0
numpy>=2.0.0
pillow>=10.0.0

# Object Detection Models (RT-DETR, D-FINE)
transformers>=4.45.0
timm>=1.0.0

# Data handling
pycocotools>=2.0.6
albumentations>=2.0.0
opencv-python-headless>=4.10.0

# Metrics
torchmetrics>=1.0.0
scipy>=1.10.0

# Supabase
supabase>=2.0.0

# HTTP
httpx>=0.27.0

# RunPod
runpod>=1.7.0

# YAML handling
pyyaml>=6.0

# Progress
tqdm>=4.65.0
